[Arun] - assignment 1.3 implemented.
What changes have you made?
findFirst implemented.it allow a particular function to call inside
findFirst.and it will find the first element that satisfy the inner
function.

What can this change impact?
it allows you to find the first element form an array that satisfy
the condition of whatever module you passes in it.